# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.ProductRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      images:
        items:
          type: string
        minItems: 1
        type: array
      name:
        maxLength: 200
        minLength: 3
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - category_id
    - description
    - images
    - name
    - price
    - stock
    type: object
...
  ecommerce-backend_internal_delivery_http.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/http.Error'
      message:
        type: string
      meta:
        $ref: '#/definitions/http.Meta'
      success:
        type: boolean
    type: object
  http.Error:
    properties:
      code:
        type: string
      details: {}
    type: object
  http.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the backend API for E-Commerce platform
  title: E-Commerce API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Browse and search products with optional filters and pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Category ID
        in: query
        name: category
        type: string
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      - description: Filter by stock availability
        in: query
        name: in_stock
        type: boolean
      - description: Search keyword
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product in the catalog. Requires admin role.
      parameters:
      - description: Product Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    delete:
      description: Delete a product. Requires admin role.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - products
    get:
      description: Get a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      summary: Get product details
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product. Requires admin role.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - products
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"